+++ js/templates/artist-view-screen.js
