+++ js/templates/change-state-game-view-screen.js
