+++ js/templates/genre-view-screen.js
