+++ js/templates/succes-view-screen.js
@@ -1,5 +1,5 @@
 import {renderTemplate} from "./../utils";
-import welcomeScreen from "./welcome-screen";
+import welcomeScreen from "./welcome-view-screen";
 import {results} from "./../data/models/endGame";
 import SuccessView from "../view/result-success-view";
