+++ js/templates/welcome-view-screen.js
